Create routes for GET
The flask framework allows the definition of functions that will be
executed when an http request is sent into a specific route with a
specific method. It also allows the definition of dynamic routes with
parameters on the url.
The routes are defined to utilize the dynamic capabilities, to allow
getting an specific resource given its id